Home work240723

+++ src/main/java/org/ru/skypro/lessons/spring/EmployeeApplication/repository/PositionRepository.java
@@ -3,7 +3,7 @@ package org.ru.skypro.lessons.spring.EmployeeApplication.repository;
 import org.ru.skypro.lessons.spring.EmployeeApplication.model.Position;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PositionRepository extends CrudRepository<Position, Integer> {
+public interface PositionRepository extends CrudRepository<Position, Long> {
     Position findByPositionName(String positionName);
 
     Position findByPositionId(Long positionId);
